Subclassed MedicalVolume to correctly copy headers
