+++ src/dcm2omids.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-from ormir_mids.dcm2omids import convert_dicom_to_ormirmids
+from ormir_mids.dcm2omids import main
 
 if __name__ == '__main__':
     main()
