+++ src/ormir_mids/converters/megre_siemens.py
@@ -3,7 +3,7 @@ import os
 import numpy as np
 
 from .abstract_converter import Converter
-from voxel import MedicalVolume
+from ..utils.OMidsMedVolume import OMidsMedVolume as MedicalVolume
 from ..utils.headers import get_raw_tag_value, group, slice_volume_3d, get_manufacturer
