+++ src/ormir_mids/converters/mese_siemens.py
@@ -1,7 +1,7 @@
 import os
 
 from .abstract_converter import Converter
-from voxel import MedicalVolume
+from ..utils.OMidsMedVolume import OMidsMedVolume as MedicalVolume
 from ..utils.headers import get_raw_tag_value, group, get_manufacturer
 
 class MeSeConverterSiemensMagnitude(Converter):
