+++ src/ormir_mids/utils/image.py
@@ -1,4 +1,4 @@
-from voxel import MedicalVolume
+from .OMidsMedVolume import OMidsMedVolume as MedicalVolume
 import numpy as np
 from scipy.ndimage import map_coordinates
